pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "foambo"
version = "0.1.2"
authors = [
  { name="Mohammed Elwardi Fadeli", email="elwardifadeli@gmail.com" },
]
description = "Configuration-based Multi-objective Bayesian Optimization on OpenFOAM cases with Ax-platform"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Environment :: GPU",
    "Framework :: Dash",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "ax-platform==0.4.1",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "PyFoam>=2023.7",
    "regex>=2023.6.3",
    "dash>=2.11.1",
    "dash_bootstrap_components>=1.4.2",
    "kaleido>=0.2.1",
]

[project.scripts]
    foamBO = "foambo.foamBO:exp_main"
    foamDash = "foambo.foamDash:dash_main"
    foamValidateBO = "foambo.foamValidate:val_main"

[project.urls]
"Homepage" = "https://github.com/FoamScience/OpenFOAM-Multi-Objective-Optimization"
"Bug Tracker" = "https://github.com/FoamScience/OpenFOAM-Multi-Objective-Optimization/issues"
